requirements.txt
dtw-python==1.1.12
keras==2.7.0
librosa==0.8.1
matplotlib==3.5.0
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.2
scipy==1.6.3
tables==3.7.0
tensorflow==2.7.0